requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=1.10.7
pydantic-settings>=2.0.0
pydantic-extra-types>=2.0.0
email-validator>=2.0.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
google-cloud-firestore>=2.11.0
google-auth>=2.17.3
python-dotenv>=1.0.0
httpx>=0.24.0
pytest>=7.3.1
pytest-asyncio>=0.21.0